fix(checkpoint): fix uint32 overflow in calcWALSize() for large page sizes
This commit addresses two issues from code review:

HIGH PRIORITY - WAL size calculation overflow:
- Fixed calcWALSize() to perform multiplication in int64 space instead
  of uint32, preventing overflow with large page sizes (>=16KB)
- Impact: With default TruncatePageN=500000:
  * 32KB pages: Was triggering at ~3.5GB instead of ~16.4GB
  * 64KB pages: Was triggering at ~2.7GB instead of ~32.8GB
- This caused the emergency TRUNCATE checkpoint to fire far earlier
  than configured, defeating the purpose of the threshold
- Added comprehensive regression test (TestCalcWALSize) with 4KB,
  16KB, 32KB, and 64KB page sizes, including overflow detection
- Created export_test.go to expose calcWALSize for testing

LOW PRIORITY - Soak test config cleanup:
- Updated tests/integration/soak_helpers.go to emit truncate-page-n
  instead of removed max-checkpoint-page-count field
- This prevents misleading config for anyone reading generated
  soak test configurations

Documentation updates:
- Removed all MaxCheckpointPageN references from documentation
- Updated checkpoint examples to use new 3-tier strategy
- Updated YAML config examples in agent documentation
- All remaining references are appropriate (documenting the removal)

+++ db.go
@@ -733,8 +733,9 @@ func (db *DB) walFileSize() (int64, error) {
 }
 
 // calcWALSize returns the size of the WAL for a given page size & count.
+// Casts to int64 before multiplication to prevent uint32 overflow with large page sizes.
 func calcWALSize(pageSize uint32, pageN uint32) int64 {
-	return int64(WALHeaderSize + ((WALFrameHeaderSize + pageSize) * pageN))
+	return int64(WALHeaderSize) + (int64(WALFrameHeaderSize+pageSize) * int64(pageN))
 }
 
 // ensureWALExists checks that the real WAL exists and has a header.
